IBM-only: Fixed smb version in debian/rules to fix U22 upgrade issue
The nfs-ganesha is depends on smb for libwbclient.so.0. When ganesha
build on ubunutu22, the debhelper generates dependency basded on
shared library installed on the system. This will work fine for fresh
scale installation.
But in case of ubuntu+ganesha upgrade, the NFS upgrade will happen first,
then smb upgrade. The problem is that, the debhelper generate build
dependency based on current smb version installed on ubuntu22, which is
latest smb version. So, NFS upgrade look for latest smb verion and fails.
it is failing in dependency check.

Fixed the issue by setting old version of smb , so that ganesha upgrade
will go through. This is just to overcome dependency failure, ganesha
doesn't rely on particular shared library version.

Change-Id: I6c5220281cf6195407706cc575cbdb7dbae9a504
Signed-off-by: Prabhu Murugesan <prabhu.murugeshan@ibm.com>