Incorrect values for ganesha_stats json option
Fixes: https://github.com/nfs-ganesha/nfs-ganesha/issues/1003

This commit fixes the second issue mentioned in the bug.

Issue: The json option of ganesha_stats utility for iov3, iov4
etc gives incorrect values for write. Due to that we cannot use
this option. Without json option the values are correct.

Fix: Incorrect element from dbus result was printed. The logic
for printing the result was not correct. Now printing the result
from correct index.

Change-Id: I639a56290649fd57ab2a8b3260798c0e5bde1c61
Signed-off-by: Trupti Shete <tshete@ddn.com>