Use after free in fd_lru
Currently release function doesn't use close_fsal_fd
instead it tries to replicate the functionality of
close_fsal_fd with the exception that it doesn't
remove the fd from fd lru.This is problematic
because if the handle has a fd associated with it
then the fd doesn't get released from the fd lru.

There are code paths in mdcache - mdcache_new_entry
for example that call release without calling
close. This leads to use after free scenarios
where the fd stays in the fd lru long after it is
freed.

We caught one such use after free scenario

TODO: This patch can be ignored in ganesha 6 as upstream patch will ready/merged

Change-Id: Iac787698783e3ded8f145c2d588a397f4f236890
Signed-off-by: Yogendra Charya Tangalapally <Yogendra.Tangalapally@ibm.com>