In nfs4_Check_Stateid don't get obj2 reference if no fsal_obj
In the call to get_state_obj_export_owner_refs, don't ask for obj2 if
fsal_obj is NULL. We don't need the obj2 reference, and we may not have
an op_context to be able to release the reference.

This is indeed the case when FREE_STATEID is called without having done
a PUTFH. Note that subsequently, FREE_STATEID DOES make it's own call
to get_state_obj_export_owner_refs, but it uses the export returned to
set up an op_context thus allowing for the put_ref.

Fixes github issue #1151

Change-Id: I71bca94c03c120265d610d43784ca879950f18bc
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>